add modal helper, optimize rename utils
