+++ packages/zowe-explorer/src/webviews/src/config-editor/components/modals/DeleteConfirmationModal.tsx
@@ -10,7 +10,7 @@
  */
 
 import * as l10n from "@vscode/l10n";
-import { useModalClickOutside, useModalFocus } from "../../hooks";
+import { ModalShell } from "../ModalShell";
 
 interface DeleteConfirmationModalProps {
   isOpen: boolean;
@@ -21,9 +21,6 @@ interface DeleteConfirmationModalProps {
 }
 
 export function DeleteConfirmationModal({ isOpen, type, name, onConfirm, onCancel }: DeleteConfirmationModalProps) {
-  const { modalRef: _clickOutsideRef, handleBackdropMouseDown, handleBackdropClick } = useModalClickOutside(onCancel);
-  const modalRef = useModalFocus(isOpen, ".delete-confirm-button");
-
   if (!isOpen) return null;
 
   const getTitle = () => {
@@ -41,51 +38,59 @@ export function DeleteConfirmationModal({ isOpen, type, name, onConfirm, onCance
   };
 
   return (
-    <div className="modal-backdrop" onMouseDown={handleBackdropMouseDown} onClick={handleBackdropClick}>
-      <div className="modal" ref={modalRef} style={{ minWidth: "400px", maxWidth: "500px" }}>
-        <div style={{ marginBottom: "16px" }}>
-          <h2 style={{ margin: "0 0 12px 0", fontSize: "18px", fontWeight: "600" }}>{getTitle()}</h2>
-          <p style={{ margin: "0", fontSize: "13px", lineHeight: "1.5", color: "var(--vscode-foreground)" }}>{getMessage()}</p>
-        </div>
-        <div style={{ display: "flex", justifyContent: "flex-end", gap: "8px", marginTop: "20px" }}>
-          <button
-            className="modal-button secondary"
-            onClick={onCancel}
-            style={{
-              padding: "8px 16px",
-              fontSize: "13px",
-              height: "32px",
-              lineHeight: "16px",
-              backgroundColor: "var(--vscode-button-secondaryBackground)",
-              color: "var(--vscode-button-secondaryForeground)",
-              border: "1px solid var(--vscode-button-secondaryBorder)",
-              borderRadius: "4px",
-              cursor: "pointer",
-              fontWeight: "normal",
-            }}
-          >
-            {l10n.t("Cancel")}
-          </button>
-          <button
-            className="modal-button delete-confirm-button"
-            onClick={onConfirm}
-            style={{
-              padding: "8px 16px",
-              fontSize: "13px",
-              height: "32px",
-              lineHeight: "16px",
-              backgroundColor: "var(--vscode-button-background)",
-              color: "var(--vscode-button-foreground)",
-              border: "1px solid var(--vscode-button-border)",
-              borderRadius: "4px",
-              cursor: "pointer",
-              fontWeight: "500",
-            }}
-          >
-            {l10n.t("Delete")}
-          </button>
-        </div>
+    <ModalShell
+      isOpen={isOpen}
+      onClose={onCancel}
+      initialFocusSelector=".delete-confirm-button"
+      titleId="delete-confirmation-modal-title"
+      overlayClassName="modal-backdrop"
+      panelClassName="modal"
+      panelStyle={{ minWidth: "400px", maxWidth: "500px" }}
+    >
+      <div style={{ marginBottom: "16px" }}>
+        <h2 id="delete-confirmation-modal-title" style={{ margin: "0 0 12px 0", fontSize: "18px", fontWeight: "600" }}>
+          {getTitle()}
+        </h2>
+        <p style={{ margin: "0", fontSize: "13px", lineHeight: "1.5", color: "var(--vscode-foreground)" }}>{getMessage()}</p>
       </div>
-    </div>
+      <div style={{ display: "flex", justifyContent: "flex-end", gap: "8px", marginTop: "20px" }}>
+        <button
+          className="modal-button secondary"
+          onClick={onCancel}
+          style={{
+            padding: "8px 16px",
+            fontSize: "13px",
+            height: "32px",
+            lineHeight: "16px",
+            backgroundColor: "var(--vscode-button-secondaryBackground)",
+            color: "var(--vscode-button-secondaryForeground)",
+            border: "1px solid var(--vscode-button-secondaryBorder)",
+            borderRadius: "4px",
+            cursor: "pointer",
+            fontWeight: "normal",
+          }}
+        >
+          {l10n.t("Cancel")}
+        </button>
+        <button
+          className="modal-button delete-confirm-button"
+          onClick={onConfirm}
+          style={{
+            padding: "8px 16px",
+            fontSize: "13px",
+            height: "32px",
+            lineHeight: "16px",
+            backgroundColor: "var(--vscode-button-background)",
+            color: "var(--vscode-button-foreground)",
+            border: "1px solid var(--vscode-button-border)",
+            borderRadius: "4px",
+            cursor: "pointer",
+            fontWeight: "500",
+          }}
+        >
+          {l10n.t("Delete")}
+        </button>
+      </div>
+    </ModalShell>
   );
 }
