+++ packages/zowe-explorer/tsconfig.json
@@ -51,6 +51,11 @@
     "src/webviews/src/config-editor/utils/generalUtils.ts",
     "src/webviews/src/config-editor/utils/renameUtils.ts",
     "src/webviews/src/config-editor/utils/profileUtils.ts",
+    "src/webviews/src/config-editor/utils/profileRenames.ts",
+    "src/webviews/src/config-editor/utils/profileTypeResolution.ts",
+    "src/webviews/src/config-editor/utils/profileSecure.ts",
+    "src/webviews/src/config-editor/utils/profileMergedProperties.ts",
+    "src/webviews/src/config-editor/utils/profileKeyListing.ts",
     "src/webviews/src/config-editor/utils/configUtils.ts",
     "src/webviews/src/config-editor/utils/profileHelpers.ts",
     "src/webviews/src/config-editor/App.tsx",
